cleaning up the code
