+++ src/landing/components/CTASection.tsx
@@ -6,12 +6,11 @@ import { motion } from "framer-motion";
 const CTASection = () => {
   return (
     <div className="py-16 sm:py-24 bg-dark-background relative">
-      {/* Background effects */}
       <div className="absolute inset-0 overflow-hidden -z-10">
         <div className="absolute top-0 left-0 w-full h-full bg-mesh-gradient opacity-50 animate-pulse-slow"></div>
-        <div className="absolute top-1/3 left-1/4 w-64 h-64 bg-purple-600/10 rounded-full blur-3xl animate-pulse-slow"></div>
+        <div className="absolute top-1/3 left-1/4 w-64 h-64 bg-blue-600/70 rounded-full blur-3xl animate-pulse-slow"></div>
         <div
-          className="absolute bottom-1/3 right-1/4 w-80 h-80 bg-indigo-600/10 rounded-full blur-3xl animate-pulse-slow"
+          className="absolute bottom-1/3 right-1/4 w-80 h-80 bg-indigo-600/70 rounded-full blur-3xl animate-pulse-slow"
           style={{ animationDelay: "1s" }}
         ></div>
       </div>
@@ -24,7 +23,6 @@ const CTASection = () => {
           viewport={{ once: true }}
           transition={{ duration: 0.5 }}
         >
-          {/* Background decorative elements */}
           <div className="absolute top-0 left-0 w-full h-full overflow-hidden">
             <motion.div
               className="absolute -top-10 -right-10 w-40 h-40 bg-white/5 rounded-full"
