+++ src/landing/components/FAQSection.tsx
@@ -1,34 +1,7 @@
 import React from "react";
 import { motion } from "framer-motion";
 import { Plus, Minus } from "lucide-react";
-
-const faqs = [
-  {
-    question: "How does EduPro AI transform my study materials?",
-    answer:
-      "EduPro AI uses advanced machine learning algorithms to analyze your uploaded materials (PDFs, images, or text) and extract key concepts. It then organizes this information into structured notes, interactive flashcards with spaced repetition scheduling, and personalized quizzes that adapt to your knowledge gaps.",
-  },
-  {
-    question: "What types of files can I upload?",
-    answer:
-      "You can upload PDFs, images of printed or handwritten notes (through our OCR technology), Word documents, PowerPoint presentations, and plain text files. Our system processes these materials and converts them into interactive study content.",
-  },
-  {
-    question: "How accurate is the AI in generating questions?",
-    answer:
-      "Our AI generates questions with a high degree of accuracy by identifying key concepts and relationships within your study materials. The system continuously improves through machine learning, becoming more accurate with usage. You can also edit or refine any AI-generated content if needed.",
-  },
-  {
-    question: "Can I collaborate with other students?",
-    answer:
-      "Yes, our Team and Enterprise plans offer collaboration features that allow you to share study materials, flashcards, and notes with classmates. You can create study groups, assign different sections to team members, and track collective progress.",
-  },
-  {
-    question: "Is my data secure and private?",
-    answer:
-      "Absolutely. We employ bank-level encryption for all uploaded content and personal information. Your study materials remain private to you unless you explicitly choose to share them. We never use your content to train our AI without explicit permission.",
-  },
-];
+import { faqs } from "../constants/faqs"; 
 
 const FAQSection = () => {
   const [activeIndex, setActiveIndex] = React.useState<number | null>(null);
@@ -39,12 +12,11 @@ const FAQSection = () => {
 
   return (
     <div className="py-16 sm:py-24 bg-dark-background relative">
-      {/* Background gradients with animation */}
       <div className="absolute inset-0 overflow-hidden -z-10">
         <div className="absolute top-0 left-0 w-full h-full bg-mesh-gradient opacity-50 animate-pulse-slow"></div>
-        <div className="absolute top-1/3 right-1/4 w-64 h-64 bg-purple-600/10 rounded-full blur-3xl animate-pulse-slow"></div>
+        <div className="absolute top-1/3 right-1/4 w-64 h-64 bg-blue-600/70 rounded-full blur-3xl animate-pulse-slow"></div>
         <div
-          className="absolute -bottom-20 -left-20 w-80 h-80 bg-indigo-600/10 rounded-full blur-3xl animate-pulse-slow"
+          className="absolute -bottom-20 -left-20 w-80 h-80 bg-indigo-600/70 rounded-full blur-3xl animate-pulse-slow"
           style={{ animationDelay: "1s" }}
         ></div>
       </div>
