+++ src/landing/constants/faqs.ts
@@ -1,0 +1,28 @@
+export const faqs = [
+    {
+      question: "How does EduPro AI transform my study materials?",
+      answer:
+        "EduPro AI uses advanced machine learning algorithms to analyze your uploaded materials (PDFs, images, or text) and extract key concepts. It then organizes this information into structured notes, interactive flashcards with spaced repetition scheduling, and personalized quizzes that adapt to your knowledge gaps.",
+    },
+    {
+      question: "What types of files can I upload?",
+      answer:
+        "You can upload PDFs, images of printed or handwritten notes (through our OCR technology), Word documents, PowerPoint presentations, and plain text files. Our system processes these materials and converts them into interactive study content.",
+    },
+    {
+      question: "How accurate is the AI in generating questions?",
+      answer:
+        "Our AI generates questions with a high degree of accuracy by identifying key concepts and relationships within your study materials. The system continuously improves through machine learning, becoming more accurate with usage. You can also edit or refine any AI-generated content if needed.",
+    },
+    {
+      question: "Can I collaborate with other students?",
+      answer:
+        "Yes, our Team and Enterprise plans offer collaboration features that allow you to share study materials, flashcards, and notes with classmates. You can create study groups, assign different sections to team members, and track collective progress.",
+    },
+    {
+      question: "Is my data secure and private?",
+      answer:
+        "Absolutely. We employ bank-level encryption for all uploaded content and personal information. Your study materials remain private to you unless you explicitly choose to share them. We never use your content to train our AI without explicit permission.",
+    },
+  ];
+
