+++ src/landing/constants/testimonials.ts
@@ -1,0 +1,31 @@
+export const testimonials = [
+    {
+      name: "Alex Esther",
+      role: "Medical Student",
+      image:
+        "https://images.unsplash.com/photo-1506794778202-cad84cf45f1d?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=687&q=80",
+      text: "Invoicer AI revolutionized how I study for my medical exams. The flashcards with spaced repetition helped me memorize complex terminology in half the time it used to take.",
+    },
+    {
+      name: "real Moses",
+      role: "Computer Science Major",
+      image:
+        "https://images.unsplash.com/photo-1534528741775-53994a69daeb?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=764&q=80",
+      text: "Being able to upload my programming lecture PDFs and have them converted into interactive quizzes was a game-changer. I identified knowledge gaps I didn't even know I had.",
+    },
+    {
+      name: "Michael Chen",
+      role: "PhD Candidate",
+      image:
+        "https://images.unsplash.com/photo-1507003211169-0a1dd7228f2d?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=687&q=80",
+      text: "The organized content management system saved me countless hours. Being able to tag my research materials by topic and easily retrieve them has streamlined my dissertation work significantly.",
+    },
+    {
+      name: "Emma Rodriguez",
+      role: "Law Student",
+      image:
+        "https://images.unsplash.com/photo-1494790108377-be9c29b29330?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=687&q=80",
+      text: "The note generation feature helped me condense lengthy case studies into concise summaries with all the key points highlighted. My study efficiency improved by at least 40%.",
+    },
+  ];
+
